# OpenAPI/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/murals/{id}/thumbnail:
    get:
      tags:
      - Murals
      summary: Get mural thumbnail
      description: Retrieve the thumbnail image of a mural
      operationId: getThumbnail
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  format: byte
        "200":
          description: Thumbnail retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  format: byte
    put:
      tags:
      - Murals
      summary: Update mural thumbnail
      description: Replace the mural's thumbnail image
      operationId: changeThumbnail
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  description: Image file
                  format: binary
      responses:
        "400":
          description: Invalid image format
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Thumbnail updated successfully
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
  /api/murals/{id}/banner:
    get:
      tags:
      - Murals
      summary: Get mural banner
      description: Retrieve the banner image of a mural
      operationId: getBanner
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  format: byte
        "200":
          description: Banner retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  type: string
                  format: byte
    put:
      tags:
      - Murals
      summary: Update mural banner
      description: Replace the mural's banner image
      operationId: changeBanner
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  description: Image file
                  format: binary
      responses:
        "204":
          description: Banner updated successfully
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Invalid image format
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
...
  /api/murals:
    get:
      tags:
      - Murals
      summary: Get murals
      description: Retrieve murals of a specific type
      operationId: getMurals
      parameters:
      - name: type
        in: query
        description: Type of murals to retrieve
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Murals retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MuralDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MuralDTO"
    post:
      tags:
      - Murals
      summary: Create new mural
      description: Create a new mural with thumbnail and banner images
      operationId: createMural
      parameters:
      - name: name
        in: query
        description: Mural name
        required: true
        schema:
          type: string
      - name: description
        in: query
        description: Mural description
        required: true
        schema:
          type: string
      - name: visibility
        in: query
        description: "Visibility type: PUBLIC or PRIVATE"
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
              - banner
              - thumbnail
              type: object
              properties:
                thumbnail:
                  type: string
                  description: Thumbnail image file
                  format: binary
                banner:
                  type: string
                  description: Banner image file
                  format: binary
      responses:
        "200":
          description: Mural created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
        "400":
          description: Invalid image format or visibility
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
  /api/murals/{muralId}/users/{userId}/unban:
    post:
      tags:
      - Murals
      summary: Unban user from mural
      description: Remove a ban for a user from a mural
      operationId: unbanUser
      parameters:
      - name: muralId
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: User ID to unban
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User unbanned successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "404":
          description: Mural or user not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "403":
          description: Forbidden - insufficient permissions
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
  /api/murals/{muralId}/users/{userId}/ban:
    post:
      tags:
      - Murals
      summary: Ban user from mural
      description: Ban a user from a mural
      operationId: banUser
      parameters:
      - name: muralId
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: User ID to ban
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural or user not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "403":
          description: Forbidden - insufficient permissions
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "200":
          description: User banned successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
  /api/murals/join:
    post:
      tags:
      - Murals
      summary: Join mural
      description: Join a mural using a code or mural ID
      operationId: joinMural
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: integer
                format: int32
        "200":
          description: Successfully joined mural
          content:
            '*/*':
              schema:
                type: integer
                format: int32
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: integer
                format: int32
...
  /api/murals/{id}:
    get:
      tags:
      - Murals
      summary: Get mural by ID
      description: Retrieve a mural's information by its ID
      operationId: getMural
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
        "200":
          description: Mural found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
    delete:
      tags:
      - Murals
      summary: Delete mural
      description: Delete a mural and all its contents
      operationId: deleteMural
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "403":
          description: Forbidden - only owner can delete
          content:
            '*/*':
              schema:
                type: string
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: string
        "204":
          description: Mural deleted successfully
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: string
    patch:
      tags:
      - Murals
      summary: Edit mural
      description: Update mural information
      operationId: editMural
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MuralEditDTO"
        required: true
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
        "200":
          description: Mural updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
        "403":
          description: Forbidden - insufficient permissions
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MuralDTO"
...
  /api/murals/{id}/isVisible:
    get:
      tags:
      - Murals
      summary: Check mural visibility
      description: Check if a mural is visible to the authenticated user
      operationId: isVisibleByAuthUser
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: boolean
        "200":
          description: Visibility status retrieved
          content:
            '*/*':
              schema:
                type: boolean
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: boolean
...
  /api/murals/{muralId}/users/{userId}:
    delete:
      tags:
      - Murals
      summary: Remove user from mural
      description: Remove a specified user from a mural
      operationId: removeSpecifiedUserFromMural
      parameters:
      - name: muralId
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: User ID to remove
        required: true
        schema:
          type: integer
          format: int64
      - name: inheritor
        in: query
        description: New owner ID if current owner is leaving
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural or user not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "200":
          description: User removed from mural
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "403":
          description: Forbidden - insufficient permissions
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
  /api/murals/{id}/users/me:
    delete:
      tags:
      - Murals
      summary: Leave mural
      description: Remove the current user from a mural
      operationId: removeCurrentUserFromMural
      parameters:
      - name: id
        in: path
        description: Mural ID
        required: true
        schema:
          type: integer
          format: int64
      - name: inheritor
        in: query
        description: New owner ID if current owner is leaving
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Mural not found
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "200":
          description: User removed from mural
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NamedId"
...
components:
  schemas:
...
    MuralDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        code:
          type: string
        visibility:
          type: string
        thumbnailUrl:
          type: string
        bannerUrl:
          type: string
        owner:
          $ref: "#/components/schemas/NamedId"
        members:
          type: array
          items:
            $ref: "#/components/schemas/NamedId"
        bannedUsers:
          type: array
          items:
            $ref: "#/components/schemas/NamedId"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/NamedId"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/NamedId"
    NamedId:
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int64
...
    MuralEditDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        newOwner:
          type: integer
          format: int64
...